+++ lib/skogsra/sys.ex
@@ -10,9 +10,9 @@ defmodule Skogsra.Sys do
   # Callback
 
   @impl Skogsra.Binding
-  def get_env(env)
+  def get_env(env, config)
 
-  def get_env(%Env{} = env) do
+  def get_env(%Env{} = env, _config) do
     module = Application.get_env(:skogsra, :system_module, System)
     name = Env.os_env(env)
     value = apply(module, :get_env, [name])