+++ test/support/test_config.ex
@@ -4,17 +4,18 @@ defmodule Skogsra.TestConfig do
   """
   use Skogsra
 
+  app_env :no_options, :my_app, :no_options
+
   app_env :from_system, :my_app, :from_system,
-    skip_config: true,
+    binding_skip: [:config],
     type: :integer
 
   app_env :from_config, :my_app, :from_config,
-    skip_system: true,
+    binding_skip: [:system],
     type: :integer
 
   app_env :from_default, :my_app, :from_default,
-    skip_system: true,
-    skip_config: true,
+    binding_skip: [:system, :config],
     default: 42
 
   app_env :cached, :my_app, :cached,
@@ -28,12 +29,18 @@ defmodule Skogsra.TestConfig do
   app_env :reloadable, :my_app, :reloadable, default: 42
 
   app_env :required, :my_app, :required,
-    skip_system: true,
+    binding_skip: [:system],
     required: true
 
   app_env :not_required, :my_app, :not_required,
-    skip_system: true,
+    binding_skip: [:system],
     required: false
 
   app_env :custom_type, :my_app, :custom_type, type: Skogsra.IntegerList
+
+  app_env :from_json, :my_app, :from_json,
+    os_env: "MY_JSON_VALUE",
+    binding_order: [:system, :config, Skogsra.JsonBinding],
+    config_path: "./test/support/fixtures/config.json",
+    default: 42
 end